requirements.txt
Django==1.5.5
pytz==2013.7            # Not "required", but should always be included when dealing with time
python-dateutil==2.1

PyMySQL==0.6.1          # Pure python mysql

Coffin==0.3.8           # Makes Jinja2 integration into django pretty seamless
Jinja2==2.7.1           # Higher performance and generally better template system
MarkupSafe==0.18        # Makes Jinja2 escaping faster

django-suit==0.2.5      # Much better looking django admin
django-widget-tweaks==1.3 # Modify form widget attrs in the template

webassets==0.9
django-assets==0.8
yuicompressor==2.4.8

dj-static==0.0.5        # WSGI app for easily serving statics in a Heroku friendly and more efficient way than django default static
django-storages==1.1.8  # Use me + boto for high traffic applications
boto==2.15.0            # Requires a S3 bucket and proper AWS config. More work, but higher performance

gunicorn==18.0
greenlet==0.4.1
./ext/gevent-1.0rc3.tar.gz
